Refactor IntentManager into smaller, focused interfaces
This commit refactors the `IntentManager` interface and its implementation into several smaller, more focused interfaces and their corresponding implementations. This improves modularity and maintainability by separating concerns related to different types of intents.

The following new interfaces and their implementations have been introduced:
- `ActivityResultIntentManager` and `ActivityResultIntentManagerImpl`
- `ApplicationDataIntentManager` and `ApplicationDataIntentManagerImpl`
- `AutofillIntentManager` and `AutofillIntentManagerImpl`
- `CredentialManagerIntentManager` and `CredentialManagerIntentManagerImpl`
- `ExternalLinkIntentManager` and `ExternalLinkIntentManagerImpl`
- `FileIntentManager` and `FileIntentManagerImpl`
- `ShareIntentManager` and `ShareIntentManagerImpl`
- `TileIntentManager` and `TileIntentManagerImpl`

The main `IntentManager` interface now delegates to these new, more specific interfaces. The `IntentManagerImpl` has been updated to accept these new managers as constructor parameters and delegate calls accordingly.

Additionally, `FileData` and `ShareData` have been moved to their own files within the `intent` package.

The `BitwardenBuildConfigManager` interface has been updated to include `applicationId` and `fileProviderAuthority` properties.
