+++ app/src/main/kotlin/com/x8bit/bitwarden/ui/platform/manager/intent/ActivityResultIntentManager.kt
@@ -1,0 +1,20 @@
+package com.x8bit.bitwarden.ui.platform.manager.intent
+
+import android.content.Intent
+import androidx.activity.compose.ManagedActivityResultLauncher
+import androidx.activity.result.ActivityResult
+import androidx.compose.runtime.Composable
+
+/**
+ * A manager interface for handling intents related to activity results.
+ */
+interface ActivityResultIntentManager {
+    /**
+     * Start an activity using the provided [Intent] and provides a callback, via [onResult], for
+     * retrieving the [ActivityResult].
+     */
+    @Composable
+    fun getActivityResultLauncher(
+        onResult: (ActivityResult) -> Unit,
+    ): ManagedActivityResultLauncher<Intent, ActivityResult>
+}
