+++ app/src/main/kotlin/com/x8bit/bitwarden/ui/platform/manager/intent/ActivityResultIntentManagerImpl.kt
@@ -1,0 +1,26 @@
+package com.x8bit.bitwarden.ui.platform.manager.intent
+
+import android.content.Intent
+import androidx.activity.compose.ManagedActivityResultLauncher
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.runtime.Composable
+
+/**
+ * Primary implementation of [ActivityResultIntentManager] for managing activity result intents.
+ */
+class ActivityResultIntentManagerImpl : ActivityResultIntentManager {
+    /**
+     * Start an activity using the provided [Intent] and provides a callback, via [onResult], for
+     * retrieving the [ActivityResult].
+     */
+    @Composable
+    override fun getActivityResultLauncher(
+        onResult: (ActivityResult) -> Unit,
+    ): ManagedActivityResultLauncher<Intent, ActivityResult> =
+        rememberLauncherForActivityResult(
+            contract = ActivityResultContracts.StartActivityForResult(),
+            onResult = onResult,
+        )
+}
