+++ app/src/main/kotlin/com/x8bit/bitwarden/ui/platform/manager/intent/ApplicationDataIntentManager.kt
@@ -1,0 +1,11 @@
+package com.x8bit.bitwarden.ui.platform.manager.intent
+
+/**
+ * A manager interface for handling intents related to application data.
+ */
+interface ApplicationDataIntentManager {
+    /**
+     * Starts the application's settings activity.
+     */
+    fun startApplicationDetailsSettingsActivity()
+}
