+++ app/src/main/kotlin/com/x8bit/bitwarden/ui/platform/manager/intent/ExternalLinkIntentManager.kt
@@ -1,0 +1,18 @@
+package com.x8bit.bitwarden.ui.platform.manager.intent
+
+import android.net.Uri
+
+/**
+ * A manager interface for handling intents related to external links.
+ */
+interface ExternalLinkIntentManager {
+    /**
+     * Start a Custom Tabs Activity using the provided [Uri].
+     */
+    fun startCustomTabsActivity(uri: Uri)
+
+    /**
+     * Start an activity to view the given [uri] in an external browser.
+     */
+    fun launchUri(uri: Uri)
+}
